Split inbox in two pages
