+++ lib/constants/icon_buttons/go_back_icon_button.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:responsive_framework/responsive_framework.dart';
+
+class GoBackIconButton extends StatelessWidget {
+  const GoBackIconButton({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return ResponsiveVisibility(
+      visible: false,
+      visibleWhen: const [Condition.smallerThan(name: TABLET)],
+      child: Builder(builder: (context) {
+        return IconButton(
+          icon: const Icon(
+            FontAwesomeIcons.chevronLeft,
+            size: 20.0,
+          ),
+          onPressed: () {
+            Navigator.pop(context);
+          },
+        );
+      }),
+    );
+  }
+}
