+++ lib/screens/private/email/mailbox_overview_screen.dart
@@ -3,6 +3,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:responsive_framework/responsive_framework.dart';
+import '../../../constants/icon_buttons/go_back_text_button.dart';
 import '../../../constants/loaders/loader_spinner_widget.dart';
 import '../../../services/localization_service.dart';
 import '../../../services/mailbox_service.dart';
@@ -27,18 +28,7 @@ class MailboxOverviewScreen extends StatelessWidget {
             padding: const EdgeInsets.all(20.0),
             child:
                 Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-              TextButton(
-                  style: TextButton.styleFrom(
-                      padding: EdgeInsets.zero,
-                      minimumSize: const Size(50, 30),
-                      tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                      alignment: Alignment.centerLeft),
-                  onPressed: () => {
-                        Navigator.pop(context),
-                      },
-                  child: Text(LocalizationService.of(context)
-                          ?.translate('go_back_link_label') ??
-                      '')),
+              const GoBackTextButton(),
               const SizedBox(height: 20),
               Text(
                   LocalizationService.of(context)
